Moved common functions to seperate file
